backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
plotly==5.17.0
oandapyV20==0.7.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.3.0 